// backend/backend/src/controller/types/article-types.ts
export interface ShoppingArticle {
	name: string;
	description: string;
	amount: number;
}
export interface RequestedNameChange {
	name: string;
	newName: string;
}
export interface RequestedDescriptionChange {
	name: string;
	description: string;
}
export interface RequestedAmountChange {
	name: string;
	amount: number;
}